pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webrtc-sip-gateway"
version = "1.0.0"
description = "Production-grade WebRTC to SIP Gateway bridging modern browser communications with legacy telecom infrastructure"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Gopala Rao Kambidi"}
]
keywords = ["webrtc", "sip", "voip", "gateway", "telephony", "ims", "srtp", "sdp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Telephony",
    "Topic :: Communications :: Internet Phone",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Framework :: AsyncIO",
]
dependencies = [
    "aiohttp>=3.9.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
    "cryptography>=42.0.0",
    "aioice>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
webrtc-sip-gateway = "src.main:main"

[project.urls]
Homepage = "https://github.com/gopalakambidi/webrtc-sip-gateway"
Repository = "https://github.com/gopalakambidi/webrtc-sip-gateway"
Issues = "https://github.com/gopalakambidi/webrtc-sip-gateway/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
